+++ functions/src/functions/image/imageUtil.ts
@@ -1,5 +1,4 @@
 import * as admin from "firebase-admin";
-
 import * as path from "path";
 import * as os from "os";
 import * as fs from "fs";
@@ -105,4 +104,73 @@ export const getToFileFullPath = (filePath, size) => {
   const thumbFileName = `${size}/${fileName}`;
   const toFileFullPath = path.join(dirName, thumbFileName);
   return toFileFullPath;
+};
+
+
+export const resizeLocal = async (fromFileFullPath, size) => {
+  const tmpResizeFile = path.join(os.tmpdir(), UUID());
+  try {
+    // resize
+    await sharp(fromFileFullPath)
+      .rotate()
+      .resize(size, size, {
+        fit: "inside",
+      })
+      .toFile(tmpResizeFile);
+
+    return tmpResizeFile;
+  } catch (e) {
+    console.log("error", e);
+  }
+  return "";
+};
+export const blendLocal = async (mapFilePath, photoPath, iconPath) => {
+  const tmpFile = path.join(os.tmpdir(), UUID()) + ".jpg";
+  try {
+    // resize
+    await sharp(mapFilePath)
+      .composite([
+        {
+          input: photoPath ,
+          blend: "over",
+          top:30,
+          left:130,
+        },
+        {
+          input: iconPath ,
+          blend: "over",
+          top:200,
+          left:270,
+        },
+
+      ])
+      .toFile(tmpFile);
+    return tmpFile;
+  } catch (e) {
+    console.log("error", e);
+  }
+  return "";
+};
+
+export const uploadFileToBucket = async (tmpFile,object) => {
+  const bucketObj = admin.storage().bucket(object.bucket);
+  //const tmpResizeFile = path.join(os.tmpdir(), UUID());
+  try {
+    // upload
+    const uuid = UUID();
+    const ret = await bucketObj.upload(tmpFile, {
+      destination: object.name,
+      metadata: {
+        contentType: object.contentType,
+        metadata: {
+          firebaseStorageDownloadTokens: uuid,
+        },
+      },
+    });
+    const file = ret[0];
+    return `https://firebasestorage.googleapis.com/v0/b/${object.bucket}/o/${encodeURIComponent(file.name)}?alt=media&token=${uuid}`;
+  } catch (e) {
+    console.log("error", e);
+  }
+  return false;
 };
